factorize sample list
