+++ py4cast/datasets/base.py
@@ -5,7 +5,6 @@ and their interfaces
 
 import datetime as dt
 import json
-from abc import abstractproperty
 from copy import deepcopy
 from dataclasses import dataclass, field, fields
 from functools import cached_property
@@ -712,6 +711,57 @@ class DatasetABC(Dataset):
             diff_stats=self.diff_stats,
             state_weights=self.state_weights,
         )
+
+    @cached_property
+    def sample_list(self):
+        """Creates the list of samples."""
+        print("Start creating samples...")
+        stats = self.stats if self.settings.standardize else None
+
+        n_inputs, n_preds, step_duration = (
+            self.settings.num_input_steps,
+            self.settings.num_pred_steps,
+            self.period.step_duration,
+        )
+
+        sample_timesteps = [
+            step_duration * step for step in range(-n_inputs + 1, n_preds + 1)
+        ]
+        all_timestamps = []
+        for date in tqdm.tqdm(self.period.date_list):
+            for term in self.period.terms_list:
+                t0 = date + dt.timedelta(hours=int(term))
+                validity_times = [
+                    t0 + dt.timedelta(hours=ts) for ts in sample_timesteps
+                ]
+                terms = [dt.timedelta(hours=int(t + term)) for t in sample_timesteps]
+
+                timestamps = Timestamps(
+                    datetime=date,
+                    terms=np.array(terms),
+                    validity_times=validity_times,
+                )
+                if self.valid_timestamp(n_inputs, timestamps):
+                    all_timestamps.append(timestamps)
+
+        samples = []
+        for ts in all_timestamps:
+            for member in self.settings.members:
+                sample = Sample(
+                    ts,
+                    self.settings,
+                    self.params,
+                    stats,
+                    self.grid,
+                    self.exists,
+                    self.get_param_tensor,
+                    member,
+                )
+                if sample.is_valid():
+                    samples.append(sample)
+
+        print(f"--> All {len(samples)} {self.period.name} samples are now defined")
+        return samples
 
     def write_list_valid_samples(self):
         print(f"Writing list of valid samples for {self.period.name} set...")
